completed TypingText componet so that it is now a reusable componet and added _variables.scss partial
