+++ src/components/Header/Header.scss
@@ -1,3 +1,5 @@
+@import '../../partials/variables';
+
 .header {
   background: #151515 url('../../assets/images/intro-bg.jpg') no-repeat center center;
   background-attachment: fixed;
@@ -11,27 +13,36 @@
 
   // }
 
+  // $xs: 576px;
+  // $sm: 576px;
+  // $md: 768px;
+  // $lg: 992px;
+  // $xl: 1200px;
+  // $xxl: 1400px;
+  
   &__introduction {
+    font-size: 2rem;
     min-height: 4rem;
     max-height: 4rem;
-    font-size: 2rem;
     
-    @media(min-width: 720px) {
-      min-height: 5rem;
-      max-height: 5rem;
-      font-size: 2rem;
+    @media(min-width: $md) {
+      min-height: 4.5rem;
+      max-height: 4.5rem;
     }
     
-    @media(min-width: 960px) {
+    @media(min-width: $lg) {
       min-height: 5.5rem;
       max-height: 5.5rem;
-      font-size: 2rem;
     }
-   
-    @media(min-width: 980px) {
+    
+    @media(min-width: $xl) {
       min-height: 6rem;
       max-height: 6rem;
-      font-size: 2rem;
+    }
+    
+    @media(min-width: $xl) {
+      min-height: 6.5rem;
+      max-height: 6.5rem;
     }
   }
