+++ src/partials/_variables.scss
@@ -1,0 +1,7 @@
+// breakpoints
+$xs: 576px;
+$sm: 576px;
+$md: 768px;
+$lg: 992px;
+$xl: 1200px;
+$xxl: 1400px;
